requirements.txt
numpy==1.23.5
opencv-python==4.7.0.72
Pillow==9.5.0
tflite-runtime==2.11.0
